better store abstraction
