+++ pg_lakehouse/src/fdw/mod.rs
@@ -1,7 +1,4 @@
-mod base;
 pub mod format;
 pub mod handler;
-pub mod local;
 pub mod object_store;
 pub mod options;
-pub mod s3;
